move to binary esptool. Run get.py/get.exe to get it!
With this step we remove all dependencies, but git. Mac and Linux come
with Python by default, so they will keep using get.py to update the
toolchain and support binaries. Windows users have an "All point and
click" installation option that requires only git and Arduino to be
downloaded.
Fingers crossed :)